backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
email-validator==2.1.0.post1
dnspython==2.4.2
cryptography==41.0.8
pillow==10.1.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.2
transformers==4.36.2
spacy==3.7.2
nltk==3.8.1
opencv-python==4.8.1.78
whois==0.9.27
python-whois==0.8.0
ssl-check==1.0.12
pycryptodome==3.19.0
hashlib-compat==1.0.1
websockets==12.0
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0